Poke master to speed up packet rate when things are busy

+++ src/ao.h
@@ -981,7 +981,7 @@ struct ao_fifo {
  */
 
 #define AO_PACKET_MAX	8
-#define AO_PACKET_SYN		0xff
+#define AO_PACKET_SYN		(uint8_t) 0xff
 
 struct ao_packet {
 	uint8_t		addr;
